Refactored homepage into Home.js
